app/requirements.txt
anyio==3.6.2
certifi==2022.12.7
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
notion-client==2.0.0
python-dotenv==0.21.1
rfc3986==1.5.0
slack-bolt==1.16.1
slack-sdk==3.19.5
sniffio==1.3.0
